Lots of reorganization
